refactor: move `git` package into `cli`
