+++ .github/workflows/widgetbook-cli.yaml
@@ -15,3 +15,7 @@ jobs:
     uses: ./.github/workflows/_package.yml
     with:
       working_directory: packages/widgetbook_cli
+      post_checkout: |
+        # Setup dummy Git user
+        git config --global user.email "john@doe.co"
+        git config --global user.name "John Doe"
