--- .github/workflows/widgetbook-git.yaml
@@ -1,21 +1,0 @@
-name: widgetbook_git
-
-on:
-  pull_request:
-    paths:
-      - packages/widgetbook_git/**
-      - .github/workflows/widgetbook-git.yaml
-      - .github/workflows/_package.yml
-  push:
-    branches:
-      - main
-
-jobs:
-  _:
-    uses: ./.github/workflows/_package.yml
-    with:
-      working_directory: packages/widgetbook_git
-      post_checkout: |
-        # Setup dummy Git user
-        git config --global user.email "you@example.com"
-        git config --global user.name "Your Name"
