+++ packages/widgetbook_cli/bin/ci_parser/local_parser.dart
@@ -1,5 +1,4 @@
-import 'package:widgetbook_git/widgetbook_git.dart';
-
+import '../git/git_dir.dart';
 import './ci_parser.dart';
 
 class LocalParser extends CiParser {
