+++ packages/widgetbook_cli/bin/git/modification.dart
@@ -1,0 +1,5 @@
+enum Modification {
+  removed,
+  changed,
+  added,
+}
