+++ packages/widgetbook_cli/bin/git/tree_entry.dart
