+++ packages/widgetbook_cli/bin/helpers/helpers.dart
@@ -1,6 +1,5 @@
 export './args_helpers.dart';
 export './exceptions.dart';
-export './modification.dart';
 export './package_helper.dart';
 export './version.dart';
 export './widgetbook_zip_encoder.dart';
