+++ packages/widgetbook_cli/bin/review/use_cases/models/changed_use_case.dart
@@ -1,6 +1,6 @@
 import 'package:freezed_annotation/freezed_annotation.dart';
 
-import '../../../helpers/modification.dart';
+import '../../../git/modification.dart';
 
 part 'changed_use_case.freezed.dart';
 part 'changed_use_case.g.dart';
