+++ packages/widgetbook_cli/pubspec.yaml
@@ -15,7 +15,8 @@ dev_dependencies:
   freezed: ^2.0.3
   json_serializable: ^6.2.0
   mocktail: ^0.3.0
- 
+  test_descriptor: ^2.0.0
+
 dependencies: 
   archive: ^3.3.0
   args: ^2.3.0
@@ -31,7 +32,6 @@ dependencies:
   pubspec: ^2.3.0
   test: ^1.12.1
   version: ^3.0.2
-  widgetbook_git: ^2.2.0-beta.9
 
 executables:
   widgetbook: cli_production
