--- packages/widgetbook_git/.gitignore
@@ -1,5 +1,0 @@
-.dart_tool
-.pub
-.packages
-packages
-pubspec.lock
