--- packages/widgetbook_git/analysis_options.yaml
@@ -1,67 +1,0 @@
-include: package:lints/recommended.yaml
-
-analyzer:
-
-linter:
-  rules:
-    - always_declare_return_types
-    - always_put_required_named_parameters_first
-    - avoid_annotating_with_dynamic
-    - avoid_bool_literals_in_conditional_expressions
-    - avoid_catches_without_on_clauses
-    - avoid_catching_errors
-    - avoid_classes_with_only_static_members
-    - avoid_double_and_int_checks
-    - avoid_field_initializers_in_const_classes
-    - avoid_implementing_value_types
-    - avoid_js_rounded_ints
-    - avoid_private_typedef_functions
-    - avoid_redundant_argument_values
-    - avoid_returning_this
-    - avoid_setters_without_getters
-    - avoid_slow_async_io
-    - avoid_types_on_closure_parameters
-    - avoid_unused_constructor_parameters
-    - avoid_void_async
-    - cancel_subscriptions
-    - cascade_invocations
-    - close_sinks
-    - comment_references
-    - directives_ordering
-    - join_return_with_assignment
-    - lines_longer_than_80_chars
-    - literal_only_boolean_expressions
-    - missing_whitespace_between_adjacent_strings
-    - no_adjacent_strings_in_list
-    - no_runtimeType_toString
-    - omit_local_variable_types
-    - one_member_abstracts
-    - only_throw_errors
-    - package_api_docs
-    - prefer_asserts_in_initializer_lists
-    - prefer_const_constructors
-    - prefer_const_constructors_in_immutables
-    - prefer_const_declarations
-    - prefer_const_literals_to_create_immutables
-    - prefer_expression_function_bodies
-    - prefer_final_locals
-    - prefer_foreach
-    - prefer_int_literals
-    - prefer_interpolation_to_compose_strings
-    - prefer_mixin
-    - prefer_relative_imports
-    - prefer_single_quotes
-    - sort_pub_dependencies
-    - test_types_in_equals
-    - throw_in_finally
-    - type_annotate_public_apis
-    - unawaited_futures
-    - unnecessary_await_in_return
-    - unnecessary_lambdas
-    - unnecessary_null_aware_assignments
-    - unnecessary_parenthesis
-    - unnecessary_statements
-    - use_is_even_rather_than_modulo
-    - use_setters_to_change_properties
-    - use_string_buffers
-    - use_to_and_as_if_applicable
