--- packages/widgetbook_git/lib/widgetbook_git.dart
@@ -1,9 +1,0 @@
-export 'src/branch_reference.dart';
-export 'src/commit.dart';
-export 'src/commit_reference.dart';
-export 'src/file_diff.dart';
-export 'src/git_dir.dart';
-export 'src/git_error.dart';
-export 'src/tag.dart';
-export 'src/top_level.dart';
-export 'src/tree_entry.dart';
