--- packages/widgetbook_git/pubspec.yaml
@@ -1,21 +1,0 @@
-name: widgetbook_git
-version: 2.2.0-beta.9
-description: >-
-  Git command line wrapper. Exposes a Git directory abstraction that makes it
-  easy to inspect and manipulate a local Git repository.
-homepage: https://www.widgetbook.io?utm_source=pub.dev&utm_medium=link&utm_campaign=widgetbook_models
-repository: https://github.com/widgetbook/widgetbook
-issue_tracker: https://github.com/widgetbook/widgetbook/issues
-documentation: https://github.com/widgetbook/widgetbook/tree/main/packages/widgetbook_git
-
-environment:
-  sdk: '>=2.17.0 <4.0.0'
-
-dependencies:
-  collection: ^1.16.0
-  path: ^1.0.0
-
-dev_dependencies:
-  lints: ^2.0.1
-  test: ^1.6.0
-  test_descriptor: ^2.0.0
